feature[TW21122]: Verify Coverity Scan Medium Level

+++ plugins/org.eclipse.osee.ats.ide/src/org/eclipse/osee/ats/ide/PortPairsOperation.java
@@ -70,7 +70,7 @@ public final class PortPairsOperation extends AbstractOperation {
       }
       logf("Porting %d pairs of workflows.", portPairs.size());
 
-      double pairPercentage = 100 / portPairs.size();
+      double pairPercentage = (double) 100 / (double) portPairs.size();
       for (Pair<String, String> pair : portPairs) {
          portPair(pair);
          monitor.worked(calculateWork(pairPercentage));
